Trait tracing_distributed::Telemetry

pub trait Telemetry {
    type Visitor: Visit;
    type TraceId: Send + Sync + Clone;
    type SpanId: Send + Sync + Clone;
    fn mk_visitor(&self) -> Self::Visitor;

    fn report_span(
        &self, 
        span: Span<Self::Visitor, Self::SpanId, Self::TraceId>
    );

    fn report_event(
        &self, 
        event: Event<Self::Visitor, Self::SpanId, Self::TraceId>
    );
}

Represents the ability to publish events and spans to some arbitrary backend.

Associated Types

type Visitor: Visit

Type used to record tracing fields.

type TraceId: Send + Sync + Clone

Globally unique identifier, uniquely identifies a trace.

type SpanId: Send + Sync + Clone

Identifies spans within a trace.

Loading content...
Required methods

fn mk_visitor(&self) -> Self::Visitor

Initialize a visitor, used to record values from spans and events as they are observed

fn report_span(&self, span: Span<Self::Visitor, Self::SpanId, Self::TraceId>)

Report a Span to this Telemetry instance's backend.

fn report_event(&self, event: Event<Self::Visitor, Self::SpanId, Self::TraceId>)

Report an Event to this Telemetry instance's backend.

Loading content...